Move directive setup to separate file
